chore: split ai rewrite core
