+++ docs/agent-audit/reasoning/2026/06/06/reduce-repository-debt.json
@@ -55,12 +55,20 @@
     "infra/cloudflare/src/service/internals/mural-viewer.js",
     "infra/cloudflare/src/service/internals/mural-workspace.js",
     "infra/cloudflare/src/service/internals/mural.js",
+    "infra/cloudflare/src/core/ai-rewrite.js",
+    "infra/cloudflare/src/core/ai-rewrite/config.js",
+    "infra/cloudflare/src/core/ai-rewrite/guardrails.js",
+    "infra/cloudflare/src/core/ai-rewrite/http.js",
+    "infra/cloudflare/src/core/ai-rewrite/prompts.js",
+    "infra/cloudflare/src/core/ai-rewrite/testing.js",
+    "infra/cloudflare/src/core/ai-rewrite/text.js",
     "lychee.toml",
     "public/components/guides/api.js",
     "public/components/guides/front-matter.js",
     "public/components/guides/guides-page.js",
     "public/components/guides/pattern-controller.js",
     "public/partials/project-tabs.html",
+    "tests/ai-rewrite-split-route-state.test.js",
     "tests/govuk-pages-render-workflow-state.test.js",
     "tests/mural-service-split-route-state.test.js",
     "tests/study-guides-route-state.test.js",
@@ -74,7 +82,8 @@
     "Extracted guides API access, front matter parsing, and pattern drawer/controller behaviour from guides-page.js.",
     "Updated route-state assertions to cover extracted modules without relying on formatter-specific import line breaks.",
     "Follow-up Lychee remediation corrected stale documentation links exposed by the narrower Git ignore rules, fixed the project notes tab route, and added narrow Lychee exclusions for placeholder-heavy Mural example transcripts and the Mustache HTML head partial.",
-    "Follow-up Mural service split kept MuralServicePart as the public route surface while moving board registry resolution, direct journal sticky writing, token refresh, viewer URL probing, and workspace/room setup into responsibility-specific modules."
+    "Follow-up Mural service split kept MuralServicePart as the public route surface while moving board registry resolution, direct journal sticky writing, token refresh, viewer URL probing, and workspace/room setup into responsibility-specific modules.",
+    "Follow-up AI rewrite split kept ai-rewrite.js as the public route and Worker entry surface while moving defaults, HTTP helpers, text sanitisation, guardrails, prompt content, and testing helpers into responsibility-specific modules under infra/cloudflare/src/core/ai-rewrite/."
   ],
   "validation": [
     {
@@ -132,6 +141,34 @@
     {
       "command": "npm run format:check",
       "status": "passed_after_mural_split"
+    },
+    {
+      "command": "node --check infra/cloudflare/src/core/ai-rewrite.js && node --check infra/cloudflare/src/core/ai-rewrite/config.js && node --check infra/cloudflare/src/core/ai-rewrite/guardrails.js && node --check infra/cloudflare/src/core/ai-rewrite/http.js && node --check infra/cloudflare/src/core/ai-rewrite/prompts.js && node --check infra/cloudflare/src/core/ai-rewrite/testing.js && node --check infra/cloudflare/src/core/ai-rewrite/text.js",
+      "status": "passed_after_ai_rewrite_split"
+    },
+    {
+      "command": "node --test tests/ai-rewrite-split-route-state.test.js",
+      "status": "passed_after_ai_rewrite_split"
+    },
+    {
+      "command": "dynamic import of infra/cloudflare/src/core/ai-rewrite.js with public export assertion",
+      "status": "passed_after_ai_rewrite_split"
+    },
+    {
+      "command": "npm test",
+      "status": "passed_180_tests_after_ai_rewrite_split"
+    },
+    {
+      "command": "npm run lint",
+      "status": "passed_with_existing_warnings_after_ai_rewrite_split"
+    },
+    {
+      "command": "npm run format:check",
+      "status": "passed_after_generated_css_regeneration_after_ai_rewrite_split"
+    },
+    {
+      "command": "npm run trace:coverage",
+      "status": "passed_after_ai_rewrite_split"
     }
   ],
   "residualRisks": [
