+++ docs/agent-audit/reasoning/2026/06/06/reduce-repository-debt.md
@@ -68,12 +68,20 @@ Narrow repository ignore rules, explicitly exclude charts from Prettier and ESLi
 - infra/cloudflare/src/service/internals/mural-viewer.js
 - infra/cloudflare/src/service/internals/mural-workspace.js
 - infra/cloudflare/src/service/internals/mural.js
+- infra/cloudflare/src/core/ai-rewrite.js
+- infra/cloudflare/src/core/ai-rewrite/config.js
+- infra/cloudflare/src/core/ai-rewrite/guardrails.js
+- infra/cloudflare/src/core/ai-rewrite/http.js
+- infra/cloudflare/src/core/ai-rewrite/prompts.js
+- infra/cloudflare/src/core/ai-rewrite/testing.js
+- infra/cloudflare/src/core/ai-rewrite/text.js
 - public/components/guides/api.js
 - public/components/guides/front-matter.js
 - public/components/guides/guides-page.js
 - public/components/guides/pattern-controller.js
 - public/partials/project-tabs.html
 - tests/govuk-pages-render-workflow-state.test.js
+- tests/ai-rewrite-split-route-state.test.js
 - tests/mural-service-split-route-state.test.js
 - tests/study-guides-route-state.test.js
 - docs/agent-audit/reasoning/2026/06/06/reduce-repository-debt.md
@@ -88,6 +96,7 @@ Narrow repository ignore rules, explicitly exclude charts from Prettier and ESLi
 - Updated route-state assertions to cover the extracted modules without depending on formatter-specific single-line imports.
 - Follow-up Lychee remediation corrected stale documentation links exposed by the narrower Git ignore rules, fixed the project notes tab route, and added narrow Lychee exclusions for placeholder-heavy Mural example transcripts and the Mustache HTML head partial.
 - Follow-up Mural service split kept `MuralServicePart` as the public route surface while moving board registry resolution, direct journal sticky writing, token refresh, viewer URL probing, and workspace/room setup into responsibility-specific modules.
+- Follow-up AI rewrite split kept `ai-rewrite.js` as the public route and Worker entry surface while moving defaults, HTTP helpers, text sanitisation, guardrails, prompt content, and testing helpers into responsibility-specific modules under `infra/cloudflare/src/core/ai-rewrite/`.
 
 ## Validation attempted
 
@@ -105,6 +114,13 @@ Narrow repository ignore rules, explicitly exclude charts from Prettier and ESLi
 - Follow-up Mural split: `npm test` — passed, 179 tests.
 - Follow-up Mural split: `npm run lint` — passed with existing warnings only.
 - Follow-up Mural split: `npm run format:check` — passed.
+- Follow-up AI rewrite split: `node --check infra/cloudflare/src/core/ai-rewrite.js && node --check infra/cloudflare/src/core/ai-rewrite/config.js && node --check infra/cloudflare/src/core/ai-rewrite/guardrails.js && node --check infra/cloudflare/src/core/ai-rewrite/http.js && node --check infra/cloudflare/src/core/ai-rewrite/prompts.js && node --check infra/cloudflare/src/core/ai-rewrite/testing.js && node --check infra/cloudflare/src/core/ai-rewrite/text.js` — passed.
+- Follow-up AI rewrite split: `node --test tests/ai-rewrite-split-route-state.test.js` — passed.
+- Follow-up AI rewrite split: dynamic import of `infra/cloudflare/src/core/ai-rewrite.js` confirmed existing public exports remain available — passed.
+- Follow-up AI rewrite split: `npm test` — passed, 180 tests.
+- Follow-up AI rewrite split: `npm run lint` — passed with existing warnings only.
+- Follow-up AI rewrite split: `npm run format:check` — passed after generated CSS was regenerated by lint.
+- Follow-up AI rewrite split: `npm run trace:coverage` — passed.
 
 ## Residual risk
