+++ infra/cloudflare/src/core/ai-rewrite.js
@@ -37,650 +37,14 @@
  * @property {any}    AI Cloudflare Workers AI binding (env.AI.run).
  */
 
-/* =========================
- * @section Configuration
- * ========================= */
+import { DEFAULTS } from "./ai-rewrite/config.js";
+import { auditForBias, neutraliseInventedMethods, neutraliseInventedQuantifiers } from "./ai-rewrite/guardrails.js";
+import { corsHeaders, json } from "./ai-rewrite/http.js";
+import { DESC_SYSTEM_PROMPT, OBJ_SYSTEM_PROMPT, rulesPromptForMode, SUGGESTION_LIBRARY } from "./ai-rewrite/prompts.js";
+import { clamp, detectPII, safeParseJSON, safeText, sanitizeRewrite } from "./ai-rewrite/text.js";
 
-/**
- * Immutable configuration defaults.
- * @constant
- * @name DEFAULTS
- * @type {Readonly<{
- *   TIMEOUT_MS:number,
- *   MAX_BODY_BYTES:number,
- *   MIN_TEXT_CHARS:number,
- *   MIN_OBJ_TEXT_CHARS:number,
- *   MAX_INPUT_CHARS:number,
- *   MAX_SUGGESTIONS:number,
- *   MAX_SUGGESTION_LEN:number,
- *   MAX_REWRITE_CHARS:number,
- *   MODEL_FALLBACK:string
- * }>}
- * @default
- * @inner
- */
-const DEFAULTS = Object.freeze({
-	TIMEOUT_MS: 10_000,
-	MAX_BODY_BYTES: 512 * 1024,
-	MIN_TEXT_CHARS: 400,
-	MIN_OBJ_TEXT_CHARS: 60,
-	MAX_INPUT_CHARS: 5000,
-	MAX_SUGGESTIONS: 8,
-	MAX_SUGGESTION_LEN: 160,
-	MAX_REWRITE_CHARS: 1800, // ~220 words
-	MODEL_FALLBACK: "@cf/meta/llama-3.1-8b-instruct"
-});
-
-/* =========================
- * @section Helper functions
- * ========================= */
-
-/**
- * Build CORS headers for the given origin based on ALLOWED_ORIGINS.
- * @function corsHeaders
- * @inner
- * @param {Env} env
- * @param {string} origin
- * @returns {Record<string,string>}
- */
-function corsHeaders(env, origin) {
-	const allowed = (env.ALLOWED_ORIGINS || "").split(",").map(s => s.trim()).filter(Boolean);
-	const h = {
-		"Access-Control-Allow-Methods": "POST, OPTIONS",
-		"Access-Control-Allow-Headers": "Content-Type, Authorization",
-		"Access-Control-Max-Age": "86400",
-		"Vary": "Origin"
-	};
-	if (origin && allowed.includes(origin)) h["Access-Control-Allow-Origin"] = origin;
-	return h;
-}
-
-/**
- * JSON response helper.
- * @function json
- * @inner
- * @param {unknown} body
- * @param {number} [status=200]
- * @param {HeadersInit} [headers]
- * @returns {Response}
- */
-function json(body, status = 200, headers = {}) {
-	const hdrs = Object.assign({ "Content-Type": "application/json" }, headers || {});
-	return new Response(JSON.stringify(body), { status, headers: hdrs });
-}
-
-/**
- * Truncate long text for logs.
- * @function safeText
- * @inner
- * @param {string} t
- * @returns {string}
- */
-function safeText(t) {
-	return t && t.length > 2048 ? t.slice(0, 2048) + "…" : t;
-}
-
-/**
- * Clamp a string to max length (safe for undefined).
- * @function clamp
- * @inner
- * @param {string} s
- * @param {number} n
- * @returns {string}
- */
-function clamp(s, n) {
-	return (s || "").slice(0, n);
-}
-
-/**
- * Safe JSON.parse returning {} on failure.
- * @function safeParseJSON
- * @inner
- * @param {string} s
- * @returns {any}
- */
-function safeParseJSON(s) {
-	try { return JSON.parse(s); } catch { return {}; }
-}
-
-/**
- * Detect potential PII patterns (email, NI, NHS).
- * @function detectPII
- * @inner
- * @param {string} text
- * @returns {boolean}
- */
-function detectPII(text) {
-	const EMAIL = /\b[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}\b/i;
-	const NI = /\b(?!BG|GB|NK|KN|TN|NT|ZZ)[A-CEGHJ-PR-TW-Z]{2}\d{6}[A-D]\b/i;
-	const NHS = /\b\d{3}\s?\d{3}\s?\d{4}\b/;
-	return EMAIL.test(text) || NI.test(text) || NHS.test(text);
-}
-
-/**
- * Sanitize rewrite for PII and whitespace/newlines.
- * @function sanitizeRewrite
- * @inner
- * @param {string} s
- * @returns {string}
- */
-function sanitizeRewrite(s) {
-	let out = String(s || "");
-	// PII redaction
-	out = out.replace(/\S+@\S+/g, "[redacted]")
-		.replace(/\b(?!BG|GB|NK|KN|TN|NT|ZZ)[A-CEGHJ-PR-TW-Z]{2}\d{6}[A-D]\b/gi, "[redacted]")
-		.replace(/\b\d{3}\s?\d{3}\s?\d{4}\b/g, "[redacted]");
-	// Normalise newlines and trim line endings
-	out = out.replace(/\r\n?/g, "\n")
-		.split("\n").map(l => l.trimEnd()).join("\n")
-		.replace(/\n{3,}/g, "\n\n") // collapse 3+ blanks to 2
-		.trim();
-	return out;
-}
-
-/* =========================
- * @section Quantifier & method neutralisers (guardrails)
- * ========================= */
-
-/**
- * Extract quantified phrases from text for comparison.
- * Looks for:
- *  - counts near user terms (e.g., "12 users", "8 participants", "at least 8 participants")
- *  - percentages (e.g., "25%")
- *  - timeframes (e.g., "by end of Q2", "within 3 months")
- * @param {string} text
- * @returns {string[]} lowercased phrases (deduplicated)
- */
-function extractQuantifiedPhrases(text) {
-	const t = String(text || "");
-	const phrases = new Set();
-
-	// counts near user terms
-	const countUser = /\b(?:at\s+least\s+|up\s+to\s+|around\s+|approximately\s+)?\d+\s+(?:users?|participants?|people|sessions?)\b/gi;
-	// generic counts (keep modest to avoid overcatch)
-	const countGeneric = /\b(?:at\s+least\s+|up\s+to\s+)?\d+\s+(?:interviews?|tests?|studies|sessions?)\b/gi;
-	const percentages = /\b\d{1,3}%\b/g;
-	const timeframeQ = /\bby\s+end\s+of\s+Q[1-4]\b/gi;
-	const timeframeWithin = /\bwithin\s+\d+\s+(?:days?|weeks?|months?|quarters?)\b/gi;
-	const timeframeIn = /\bin\s+\d+\s+(?:days?|weeks?|months?|quarters?)\b/gi;
-
-	const addMatches = (re) => {
-		let m;
-		while ((m = re.exec(t)) !== null) phrases.add(m[0].toLowerCase());
-	};
-
-	[countUser, countGeneric, percentages, timeframeQ, timeframeWithin, timeframeIn].forEach(addMatches);
-
-	return Array.from(phrases);
-}
-
-/**
- * Remove/neutralise quantifiers in rewrite that do not exist in the input.
- * Also generate suggestion items explaining what was removed and why.
- *
- * Neutralisations:
- *  - "by end of Q2", "within 3 months", "in 6 weeks" -> removed (timeline becomes open).
- *  - "at least 8 participants", "12 users" -> drop the number (e.g., "participants", "users").
- *  - "25%" or "by 25%" -> replace with "a measurable amount" to avoid invented specifics.
- *
- * @param {string} rewrite
- * @param {string} input
- * @returns {{text:string, notes:Array<{category:string, tip:string, why:string, severity:"high"|"medium"|"low"}>}}
- */
-function neutraliseInventedQuantifiers(rewrite, input) {
-	let out = String(rewrite || "");
-	const inputPhrases = new Set(extractQuantifiedPhrases(input).map(s => s.toLowerCase()));
-
-	/** @type {Array<{category:string, tip:string, why:string, severity:"high"|"medium"|"low"}>} */
-	const notes = [];
-
-	// Helpers for safe replace across the whole string
-	const replaceAll = (re, replacer) => { out = out.replace(re, replacer); };
-
-	// 1) Timeframe phrases
-	const timeframePatterns = [
-		{ re: /\bby\s+end\s+of\s+Q[1-4]\b/gi, label: "Timeline" },
-		{ re: /\bwithin\s+\d+\s+(days?|weeks?|months?|quarters?)\b/gi, label: "Timeline" },
-		{ re: /\bin\s+\d+\s+(days?|weeks?|months?|quarters?)\b/gi, label: "Timeline" }
-	];
-
-	timeframePatterns.forEach(({ re, label }) => {
-		replaceAll(re, (m) => {
-			if (!inputPhrases.has(m.toLowerCase())) {
-				notes.push({
-					category: label,
-					tip: `Remove invented timeframe ('${m}').`,
-					why: "Timeframes must come from the input; propose them in suggestions instead.",
-					severity: "high"
-				});
-				return "";
-			}
-			return m;
-		});
-	});
-
-	// 2) Counts near user terms (drop numbers if invented)
-	const countUser = /\b(?:(at\s+least|up\s+to|around|approximately)\s+)?(\d+)\s+(users?|participants?|people|sessions?)\b/gi;
-	replaceAll(countUser, (m, qualifier, num, noun) => {
-		if (inputPhrases.has(m.toLowerCase())) return m; // allowed (present in input)
-		notes.push({
-			category: "Measurability",
-			tip: `Remove invented count ('${m}'); keep role only.`,
-			why: "Sample sizes must come from the input; propose targets in suggestions instead.",
-			severity: "high"
-		});
-		return noun;
-	});
-
-	// 3) Generic counts (interviews/tests/etc.)
-	const countGeneric = /\b(?:(at\s+least|up\s+to|around|approximately)\s+)?(\d+)\s+(interviews?|tests?|studies|sessions?)\b/gi;
-	replaceAll(countGeneric, (m, qualifier, num, noun) => {
-		if (inputPhrases.has(m.toLowerCase())) return m;
-		notes.push({
-			category: "Measurability",
-			tip: `Remove invented count ('${m}'); keep the activity only.`,
-			why: "Counts must come from the input; suggest them separately if useful.",
-			severity: "high"
-		});
-		return noun;
-	});
-
-	// 4) Percentages (e.g., 25%)
-	const percent = /\b(\d{1,3})%\b/g;
-	replaceAll(percent, (m) => {
-		if (inputPhrases.has(m.toLowerCase())) return m;
-		notes.push({
-			category: "Outcomes & measures",
-			tip: `Neutralise invented percentage ('${m}').`,
-			why: "Percent targets must come from the input; propose a number in suggestions instead.",
-			severity: "high"
-		});
-		return "a measurable amount";
-	});
-
-	// Clean up double spaces and stray punctuation after removals
-	out = out.replace(/[ \t]+/g, " ")
-		.replace(/\s+([,.;:])/g, "$1")
-		.replace(/\(\s*\)/g, "")
-		.replace(/\s{2,}/g, " ")
-		.replace(/\n[ \t]+/g, "\n")
-		.trim();
-
-	return { text: out, notes };
-}
-
-/**
- * Remove invented method specifics not present in the input (e.g., "using screen readers").
- * Add more phrases as needed; only strips when the phrase isn't in the input.
- */
-function neutraliseInventedMethods(rewrite, input) {
-	let out = String(rewrite || "");
-	const lcInput = String(input || "").toLowerCase();
-
-	/** Phrases to strip if absent from input */
-	const methodPhrases = [
-		"using screen readers",
-		"remote unmoderated",
-		"a/b test",
-		"benchmark test",
-		"tree test",
-		"card sort",
-		"eye tracking",
-		"heatmap",
-		"heuristic review"
-	];
-
-	for (const phrase of methodPhrases) {
-		if (!lcInput.includes(phrase)) {
-			const re = new RegExp("\\b" + phrase.replace(/[.*+?^${}()|[\]\\]/g, "\\$&") + "\\b", "gi");
-			out = out.replace(re, "");
-		}
-	}
-
-	// Tidy grammar after removals
-	out = out.replace(/[ \t]{2,}/g, " ")
-		.replace(/\s+([,.;:])/g, "$1")
-		.replace(/\(\s*\)/g, "")
-		.replace(/\s{2,}/g, " ")
-		.replace(/\s+\n/g, "\n")
-		.trim();
-
-	return out;
-}
-
-/* =========================
- * @section Bias audit + light neutralisation (guardrails)
- * ========================= */
-
-/**
- * Bias audit + light neutralisation.
- * - Scans the rewrite for common biased/excluding language.
- * - Replaces a few terms with neutral alternatives (where safe).
- * - Produces suggestion items explaining issues and mitigations.
- *
- * NOTE: We never add new project details. We only neutralise phrasing.
- *
- * @param {string} rewrite
- * @param {string} input
- * @returns {{ text: string, issues: Array<{category:string, tip:string, why:string, severity:"high"|"medium"|"low"}> }}
- */
-function auditForBias(rewrite, input) {
-	let out = String(rewrite || "");
-	/** @type {Array<{category:string, tip:string, why:string, severity:"high"|"medium"|"low"}>} */
-	const issues = [];
-
-	// --- Utility helpers ---
-	const push = (tip, why, severity = "medium", category = "Bias") =>
-		issues.push({ category, tip, why, severity });
-
-	const replaceAll = (pairs) => {
-		for (const { re, to, tip, why, severity = "medium" } of pairs) {
-			if (re.test(out)) {
-				out = out.replace(re, to);
-				push(tip, why, severity);
-			}
-		}
-	};
-
-	// --- 1) Ableist / excluding terms (safe neutralisations) ---
-	replaceAll([{
-			re: /\bwheelchair[-\s]?bound\b/gi,
-			to: "wheelchair user",
-			tip: "Use 'wheelchair user' instead of 'wheelchair-bound'.",
-			why: "‘Bound’ is ableist; the neutral term is ‘wheelchair user’.",
-			severity: "high",
-		},
-		{
-			re: /\bhandicapped\b/gi,
-			to: "disabled people",
-			tip: "Use 'disabled people' instead of 'handicapped'.",
-			why: "Aligns with inclusive, accepted language in UK public sector.",
-			severity: "high",
-		},
-		{
-			re: /\b(the\s+disabled|the\s+blind|the\s+deaf)\b/gi,
-			to: "disabled people",
-			tip: "Avoid 'the disabled' → use 'disabled people'.",
-			why: "Person-first or identity-first phrasing avoids othering.",
-			severity: "medium",
-		},
-		{
-			re: /\b(insane|crazy|mad|lame)\b/gi,
-			to: "inappropriate",
-			tip: "Avoid ableist descriptors like 'crazy' or 'lame'.",
-			why: "These terms stigmatise disability; use neutral language.",
-			severity: "medium",
-		},
-		{
-			re: /\bblind\s+spot\b/gi,
-			to: "gap",
-			tip: "Replace 'blind spot' with 'gap'.",
-			why: "Avoids ableist metaphor; neutral alternative is clearer.",
-			severity: "low",
-		},
-	]);
-
-	// --- 2) Gendered / casual collective terms ---
-	replaceAll([{
-			re: /\bchairman\b/gi,
-			to: "chair",
-			tip: "Use 'chair' instead of 'chairman'.",
-			why: "Gender-neutral job titles are more inclusive.",
-			severity: "medium",
-		},
-		{
-			re: /\bpolicemen\b/gi,
-			to: "police officers",
-			tip: "Use 'police officers' instead of 'policemen'.",
-			why: "Gender-neutral roles are clearer and inclusive.",
-			severity: "medium",
-		},
-		{
-			re: /\bguys\b/gi,
-			to: "everyone",
-			tip: "Avoid 'guys' for mixed groups; use 'everyone'.",
-			why: "Casual gendered collective can exclude readers.",
-			severity: "low",
-		},
-	]);
-
-	// --- 3) Othering / immigration & language phrasing ---
-	replaceAll([{
-			re: /\bforeigners\b/gi,
-			to: "people from outside the UK",
-			tip: "Avoid 'foreigners' → use neutral phrasing.",
-			why: "Reduces othering; focuses on context not identity.",
-			severity: "high",
-		},
-		{
-			re: /\bnon[-\s]?english\s+speakers\b/gi,
-			to: "people who do not speak English fluently",
-			tip: "Prefer 'people who do not speak English fluently'.",
-			why: "People-centred phrasing is more respectful and clear.",
-			severity: "medium",
-		},
-	]);
-
-	// --- 4) “Normal users” / defaulting language ---
-	replaceAll([{
-			re: /\bnormal\s+users?\b/gi,
-			to: "users",
-			tip: "Avoid 'normal users' → just say 'users'.",
-			why: "‘Normal’ implies others are abnormal/excluded.",
-			severity: "high",
-		},
-		{
-			re: /\bregular\s+users?\b/gi,
-			to: "users",
-			tip: "Avoid 'regular users' unless defined.",
-			why: "Imprecise category can exclude; define segments instead.",
-			severity: "low",
-		},
-	]);
-
-	// --- 5) Confirmation-bias language in research framing ---
-	// We don’t change semantics; we de-bias phrasing.
-	replaceAll([{
-			re: /\bprove\s+that\b/gi,
-			to: "test whether",
-			tip: "Replace 'prove that' with 'test whether'.",
-			why: "Reduces confirmation bias in research framing.",
-			severity: "medium",
-		},
-		{
-			re: /\bconfirm\s+that\s+users\b/gi,
-			to: "learn whether users",
-			tip: "Replace 'confirm that users' with 'learn whether users'.",
-			why: "Keeps inquiry open and avoids pre-judging outcomes.",
-			severity: "medium",
-		},
-	]);
-
-	// --- 6) “Just”, “simple”, “obvious” — trivialising difficulty ---
-	// We avoid auto-replacing these everywhere; we flag them for suggestions.
-	const trivialising = /\b(just|simply|obvious|obviously|easy|trivial)\b/gi;
-	if (trivialising.test(out)) {
-		push(
-			"Avoid words like 'just', 'simply', or 'obvious'.",
-			"They can minimise user challenges and bias facilitator expectations.",
-			"low"
-		);
-	}
-
-	// --- 7) Heuristic: testing mentioned but no accessibility mention (OK to SUGGEST only)
-	// Do not inject new content into rewrite; suggestion only.
-	const mentionsTesting = /\b(test|validate|usability|research|study|evaluate|assess)\b/i.test(out);
-	const mentionsA11y =
-		/\b(accessibility|screen\s*reader|keyboard|contrast|assistive\s+tech|WCAG|disabled\s+people)\b/i.test(out);
-	if (mentionsTesting && !mentionsA11y) {
-		// Only add if the INPUT itself doesn’t already mention it (to avoid nagging duplicates)
-		const inputHasA11y =
-			/\b(accessibility|screen\s*reader|keyboard|contrast|assistive\s+tech|WCAG|disabled\s+people)\b/i.test(input || "");
-		if (!inputHasA11y) {
-			push(
-				"Consider how accessibility needs will be included in the research.",
-				"Ensures the work is inclusive across disabilities, devices, and contexts.",
-				"medium"
-			);
-		}
-	}
-
-	// Normalise whitespace after replacements
-	out = out.replace(/[ \t]{2,}/g, " ")
-		.replace(/\s+([,.;:])/g, "$1")
-		.replace(/\s{2,}/g, " ")
-		.replace(/\n[ \t]+/g, "\n")
-		.trim();
-
-	return { text: out, issues };
-}
-
-/* =========================
- * @section Shared prompts (exported for reuse)
- * ========================= */
-
-/**
- * High-level base instruction shared by both Description and Objectives modes.
- * @constant
- * @name BASE_SYSTEM_PROMPT
- * @type {string}
- */
-export const BASE_SYSTEM_PROMPT = [
-	"You assist UK Home Office user researchers.",
-	"Use GOV.UK style: plain English, short sentences, accessible to all.",
-	"Only use facts from the provided input; never invent new details.",
-	"If personal data appears, do not repeat it; instead advise removal.",
-	"Output must be strictly JSON. Do not include markdown, code fences, or explanatory prose.",
-	"If any field would be empty, return an empty string — never omit required keys."
-].join(" ");
-
-/**
- * Description-specific system prompt (Step 1).
- * @constant
- * @name DESC_SYSTEM_PROMPT
- * @type {string}
- */
-export const DESC_SYSTEM_PROMPT = [
-	BASE_SYSTEM_PROMPT,
-	"Rewrite a research project Description.",
-	"Structure the rewrite into labelled sections only if the input supports them.",
-	"Section format (mandatory): Each section must use exactly this layout: 1). Line 1 » Section label followed by a colon, with no text after the colon. 2). Carriage return. 3). Line 2+ » Content for that section. 4). 2 Carriage returns after the section and before the next label. 5). Never place content on the same line as the label. Always 2 carriage returns between section content and the following label. Do not include unused labels.",
-	"Typical sections you may include: Problem, Scope, Users, Outcomes, Ethics, Method, Assumptions & Risks, Context, Stakeholders, Research Questions, Timeline, Recruitment, Data Handling, Success Criteria."
-].join(" ");
-
-/**
- * Objectives-specific system prompt (Step 2).
- * @constant
- * @name OBJ_SYSTEM_PROMPT
- * @type {string}
- */
-export const OBJ_SYSTEM_PROMPT = [
-	BASE_SYSTEM_PROMPT,
-	"Rewrite and refine 'Initial Objectives' for a new research project.",
-	"Group into clear, numbered objectives where possible.",
-	"Apply SMART where detail exists (specific, measurable, achievable, relevant, time-bound).",
-	"Avoid adding brand-new aims; only clarify or tighten what is present."
-].join(" ");
-
-/* =========================
- * @section Mode rules
- * ========================= */
-
-/**
- * Build the RULES prompt per mode.
- * @function rulesPromptForMode
- * @param {"description"|"objectives"} mode
- * @returns {string}
- */
-function rulesPromptForMode(mode) {
-	if (mode === "description") {
-		return [
-			"Rules (Description):",
-			"01) Problem framing: restate as a user need; add one line each for in-scope/out-of-scope if the input mentions them.",
-			"02) Users & inclusion: name primary users and contexts; mention inclusion (accessibility, device, language) if present.",
-			"03) Outcomes & measures: include SMART outcomes with a number and timeframe where available.",
-			"04) Assumptions & risks: capture as hypotheses; note constraints or dependencies.",
-			"05) Ethics: summarise consent, retention, DPIA/DPS; remove or flag PII.",
-			"06) Method: fit to maturity (discovery vs alpha) if described.",
-			"07) Context: include policy drivers, service phase, organisational context if described.",
-			"08) Stakeholders: list key people or teams to involve if given.",
-			"09) Research questions: capture explicit questions the project will address.",
-			"10) Artefacts/Deliverables: outputs such as maps, prototypes, reports if stated.",
-			"11) Timeline: milestones or expected timeframe if available.",
-			"12) Recruitment: sample, accessibility needs, demographics if mentioned.",
-			"13) Data handling: storage, retention, sharing rules if described.",
-			"14) Success criteria: capture what 'good' looks like if stated.",
-			"15) Style: expand acronyms; use plain English; short sentences.",
-			"16) Clarity: remove duplication; structure content under clear headings.",
-			"17) Do NOT add quantities, dates or policy assertions unless provided; place them in suggestions instead.",
-			"",
-			"Include only sections where the input contains relevant content. Never invent details."
-		].join("\n");
-	}
-
-	if (mode === "objectives") {
-		return [
-			"Rules (Objectives):",
-			"01) Split into 3–6 concise, numbered objectives when possible.",
-			"02) Make each objective action-oriented (start with a verb).",
-			"03) Apply SMART only if numbers or timeframes are explicitly present in the input. Do not create new metrics, percentages, or deadlines.",
-			"04) Include any constraints, dependencies, or risks if mentioned.",
-			"05) Keep scope aligned to the service phase and project status if present in the input.",
-			"06) Avoid PII; if present, advise removal in suggestions.",
-			"07) Use GOV.UK style: plain English, short sentences, expanded acronyms.",
-			"08) If there is ambiguity, keep the objective clear but do not invent details.",
-			"09) Never introduce new numbers, dates, or timeframes in the rewrite. Preserve exactly what is in the input.",
-			"10) Do NOT add numbers, sample sizes, percentages, or timeframes unless they appear in the input; propose them only in suggestions.",
-			"",
-			"The rewrite must never introduce new numbers, dates, or timeframes. Only preserve or clarify what is already present in the input."
-		].join("\n");
-	}
-
-	// Fallback if an unknown mode sneaks through
-	return "No rules available for this mode.";
-}
-
-/* =========================
- * @section Suggestion library (examples to guide AI)
- * ========================= */
-
-/**
- * Curated suggestion patterns derived from common issues.
- * The model should select only relevant items for the current input.
- * @constant
- * @name SUGGESTION_LIBRARY
- * @type {string}
- */
-const SUGGESTION_LIBRARY = [
-	"Suggestion patterns (use only if relevant to the input):",
-	"",
-	"Weak / unstructured objectives:",
-	"• Measurability: Add numeric targets to 2 objectives. — Enables progress tracking. (high)",
-	"• Clarity: Start each objective with an action verb. — Improves readability. (medium)",
-	"• Scope: Remove ambiguous objectives. — Prevents confusion. (low)",
-	"• Users & inclusion: Consider users with disabilities when testing the prototype. — Ensures accessibility. (medium)",
-	"• Outcomes & measures: Define what success looks like for each objective. — Provides direction. (high)",
-	"• Risks: Identify potential risks and mitigation strategies. — Prepares for challenges. (low)",
-	"",
-	"Overly broad aims:",
-	"• Scope: Replace 'test everything' with focused aspects (e.g., security, user satisfaction). — Enables targeted research. (high)",
-	"• Stakeholders: Clarify whose expectations will be tested and how they will be involved. — Aligns delivery. (medium)",
-	"",
-	"With measurable outcomes present:",
-	"• Measurability: Check each objective has a numeric target and timeframe. — Supports tracking. (high)",
-	"• Clarity: Avoid vague phrases like 'at least' when a precise number exists. — Improves specificity. (low)",
-	"• Inclusion: Prefer 'relevant stakeholders' over specific roles unless needed. — Improves inclusivity. (medium)",
-	"",
-	"With risks / dependencies:",
-	"• Risks: Surface low-availability recruitment risks and how to mitigate them. — Enables contingency planning. (medium)",
-	"• Constraints: Note browser/device constraints as temporary and avoid narrowing scope unless necessary. — Prevents bias. (low)",
-	"• Measurability: Add numeric targets where missing. — Supports tracking. (high)",
-	"",
-	"Mixed quality lists:",
-	"• Scope: Remove objectives that are too vague; merge overlaps. — Improves focus. (high)",
-	"• Measurability: Add measurable targets to 2 objectives. — Enables progress tracking. (high)",
-	"• Inclusion: Add accessibility-related testing where relevant (e.g., screen readers). — Ensures accessibility. (medium)"
-].join("\n");
+export { BASE_SYSTEM_PROMPT, DESC_SYSTEM_PROMPT, OBJ_SYSTEM_PROMPT } from "./ai-rewrite/prompts.js";
+export { createMockEnv, makeJsonRequest } from "./ai-rewrite/testing.js";
 
 /* =========================
  * @section Service
@@ -1026,44 +390,3 @@ export default {
 		return json({ error: "Not found" }, 404, corsHeaders(env, origin));
 	}
 };
-
-/* =========================
- * @section Test utilities (named exports)
- * ========================= */
-
-/**
- * Create a minimal mock Env for unit tests.
- * @function createMockEnv
- * @param {Partial<Env>} overrides
- * @returns {Env}
- */
-export function createMockEnv(overrides = {}) {
-	return /** @type {Env} */ ({
-		ALLOWED_ORIGINS: "https://researchops.pages.dev, https://rops-api.example.workers.dev",
-		AUDIT: "false",
-		AIRTABLE_BASE_ID: "app_base",
-		AIRTABLE_API_KEY: "key",
-		AIRTABLE_TABLE_AI_LOG: "AI_Usage",
-		MODEL: "@cf/meta/llama-3.1-8b-instruct",
-		AI: { run: async () => JSON.stringify({ summary: "ok", suggestions: [], rewrite: "example" }) },
-		...overrides
-	});
-}
-
-/**
- * Build a JSON Request for tests.
- * @function makeJsonRequest
- * @example
- * const req = makeJsonRequest("/api/ai-rewrite", { mode:"description", text: "x".repeat(420) });
- */
-export function makeJsonRequest(path, body, init = {}) {
-	const reqInit = {
-		method: "POST",
-		headers: Object.assign({ "Content-Type": "application/json" }, init.headers || {}),
-		body: JSON.stringify(body)
-	};
-	for (const k in init) {
-		if (k !== "headers") reqInit[k] = init[k];
-	}
-	return new Request(`https://example.test${path}`, reqInit);
-}
